pyproject.toml
[project]
name = "dlinktranslator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "ollama>=0.5.3",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
    "websockets>=15.0.1",
]
